# cdn-server/swagger.yaml
openapi: 3.0.3
info:
  title: WCDN API
  description: |
    The WCDN (Walrus Content Delivery Network) API provides access to decentralized storage and CDN services.

    This API follows Stripe-style REST principles with:
    - Consistent resource structures
    - Standardized error handling  
    - Cursor-based pagination
    - Bearer token authentication
  version: 1.0.0
  contact:
    name: WCDN Support
    email: support@wcdn.dev
    url: https://docs.wcdn.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.wcdn.dev/v1
    description: Production server
  - url: http://localhost:4500/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'API key obtained from the WCDN dashboard'
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Number of results to return (1-100)
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    StartingAfterParam:
      name: starting_after
      in: query
      description: Cursor for pagination (return results after this ID)
      required: false
      schema:
        type: string
    EndingBeforeParam:
      name: ending_before
      in: query
      description: Cursor for pagination (return results before this ID)
      required: false
      schema:
        type: string
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - type
            - message
          properties:
            type:
              type: string
              enum:
                - validation_error
                - authentication_error
                - permission_error
                - not_found_error
                - rate_limit_error
                - api_error
                - network_error
            message:
              type: string
            code:
              type: string
            param:
              type: string
      example:
        error:
          type: validation_error
          message: Invalid blob ID format
          code: VALIDATION_FAILED
          param: id
    PaginatedList:
      type: object
      required:
        - object
        - data
        - has_more
        - url
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items: {}
        has_more:
          type: boolean
        url:
          type: string
...
    Upload:
      type: object
      required:
        - id
        - object
        - created
        - filename
        - size
        - content_type
        - blob_id
        - status
      properties:
        id:
          type: string
          description: Unique upload identifier
          example: upload_1720000000_abc123
        object:
          type: string
          enum: [upload]
        created:
          type: integer
          format: int64
          description: Unix timestamp of creation
          example: 1720000000
        filename:
          type: string
          description: Original filename
          example: document.pdf
        size:
          type: integer
          description: File size in bytes
          example: 2048576
        content_type:
          type: string
          description: MIME type
          example: application/pdf
        blob_id:
          type: string
          description: Resulting blob ID after upload
          example: GvonK6tzar1onhLJwdPz8Q8MQY6Nx17UZQw3UmL1i8c
        status:
          type: string
          enum: [processing, completed, failed]
          description: Upload status
        vault_id:
          type: string
          description: Tusky vault ID (if applicable)
          example: vault_abc123
        parent_id:
          type: string
          description: Parent folder ID (if applicable)
...
paths:
...
  /uploads:
    post:
      summary: Create an upload
      description: Upload a file to the Walrus network
      tags: [Uploads]
      parameters:
        - name: vault_id
          in: query
          description: Tusky vault ID (optional)
          required: false
          schema:
            type: string
        - name: parent_id
          in: query
          description: Parent folder ID (optional)
          required: false
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload (max 10MB)
              required:
                - file
      responses:
        '200':
          description: Upload created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Upload'
        '400':
          description: Invalid request or file too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List uploads
      description: Retrieve a list of uploads
      tags: [Uploads]
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StartingAfterParam'
        - $ref: '#/components/parameters/EndingBeforeParam'
        - name: vault_id
          in: query
          description: Filter by vault ID
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filter by upload status
          required: false
          schema:
            type: string
            enum: [processing, completed, failed]
      responses:
        '200':
          description: List of uploads
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedList'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Upload'
...